requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cleanup==8.1.0
django-environ==0.11.2
install==1.3.5
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
